Better error handling
With this change, we will pass servers that encounter an error while doing SSH.
It could be DNS, connectivity, and that would affect the whole run.
It's much more valuable to pass, and report the error to the user. The current
behavior (the exception) stop the run, and does not report which server caused
the exception.